Fix : socket communication is migrated from game.js and emitter.js into a single webSocket.js file.
