+++ public/js/game.js
@@ -1,2 +1,2 @@
 /*! For license information please see game.js.LICENSE.txt */
-!function(e){var n={};function r(t){if(n[t])return n[t].exports;var a=n[t]={i:t,l:!1,exports:{}};return e[t].call(a.exports,a,a.exports,r),a.l=!0,a.exports}r.m=e,r.c=n,r.d=function(e,n,t){r.o(e,n)||Object.defineProperty(e,n,{enumerable:!0,get:t})},r.r=function(e){"undefined"!=typeof Symbol&&Symbol.toStringTag&&Object.defineProperty(e,Symbol.toStringTag,{value:"Module"}),Object.defineProperty(e,"__esModule",{value:!0})},r.t=function(e,n){if(1&n&&(e=r(e)),8&n)return e;if(4&n&&"object"==typeof e&&e&&e.__esModule)return e;var t=Object.create(null);if(r.r(t),Object.defineProperty(t,"default",{enumerable:!0,value:e}),2&n&&"string"!=typeof e)for(var a in e)r.d(t,a,function(n){return e[n]}.bind(null,a));return t},r.n=function(e){var n=e&&e.__esModule?function(){return e.default}:function(){return e};return r.d(n,"a",n),n},r.o=function(e,n){return Object.prototype.hasOwnProperty.call(e,n)},r.p="",r(r.s=0)}({"./src/js/canvasManager.js":function(module,__webpack_exports__,__webpack_require__){"use strict";eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var _emitter_js__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ./emitter.js */ \"./src/js/emitter.js\");\n //Create canvas\n\nvar canvas = document.getElementById('myCanvas');\nvar context = canvas.getContext('2d');\nvar mouse = {\n  x: 0,\n  y: 0,\n  width: 1,\n  color: \"black\"\n}; //Set background\n\ncontext.fillStyle = \"white\";\ncontext.fillRect(0, 0, 700, 500); //Canvas controls behaviour\n\neraseAll.addEventListener('click', function (evt) {\n  evt.preventDefault();\n  eraseDrawing();\n  _emitter_js__WEBPACK_IMPORTED_MODULE_0__[\"default\"].emitEraseAll();\n});\n\nvar eraseDrawing = function eraseDrawing() {\n  context.clearRect(0, 0, canvas.width, canvas.height);\n};\n\npencil.addEventListener('click', function (evt) {\n  evt.preventDefault();\n  mouse.width = 1;\n  mouse.color = \"black\";\n});\neraser.addEventListener('click', function (evt) {\n  evt.preventDefault();\n  mouse.width = 100;\n  mouse.color = \"white\";\n}); // canvas behaviour\n//drawing start\n\ncanvas.addEventListener('mousedown', function (evt) {\n  evt.preventDefault();\n  getMousePos(evt);\n  startDrawing(mouse);\n  _emitter_js__WEBPACK_IMPORTED_MODULE_0__[\"default\"].emitDrawingStart(mouse);\n  startTracing();\n});\n\nvar startDrawing = function startDrawing(currentMouse) {\n  context.lineWidth = currentMouse.width;\n  context.strokeStyle = currentMouse.color;\n  context.beginPath();\n  context.moveTo(currentMouse.x, currentMouse.y);\n}; // drawing tracing\n\n\nvar startTracing = function startTracing() {\n  canvas.addEventListener('mousemove', mouseMove, false);\n};\n\nvar mouseMove = function mouseMove(evt) {\n  getMousePos(evt);\n  trace(mouse);\n  _emitter_js__WEBPACK_IMPORTED_MODULE_0__[\"default\"].emitDrawingTrace(mouse);\n};\n\nvar trace = function trace(currentMouse) {\n  context.lineTo(currentMouse.x, currentMouse.y);\n  context.stroke();\n}; //stopping drawing\n\n\ncanvas.addEventListener('mouseup', function (evt) {\n  stopDrawing();\n}, false);\n\nvar stopDrawing = function stopDrawing() {\n  canvas.removeEventListener('mousemove', mouseMove, false);\n};\n\nvar getMousePos = function getMousePos(evt) {\n  var rect = canvas.getBoundingClientRect();\n  mouse.x = evt.clientX - rect.left;\n  mouse.y = evt.clientY - rect.top;\n};\n\n/* harmony default export */ __webpack_exports__[\"default\"] = ({\n  startDrawing: startDrawing,\n  trace: trace,\n  stopDrawing: stopDrawing,\n  eraseDrawing: eraseDrawing\n});\n\n//# sourceURL=webpack:///./src/js/canvasManager.js?")},"./src/js/connection.js":function(module,__webpack_exports__,__webpack_require__){"use strict";eval('__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var _emitter_js__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ./emitter.js */ "./src/js/emitter.js");\n //Get the pseudo of user\n\nvar pseudo = prompt("Entrer votre pseudonyme:", "");\n_emitter_js__WEBPACK_IMPORTED_MODULE_0__["default"].emitConnection(pseudo);\n\n//# sourceURL=webpack:///./src/js/connection.js?')},"./src/js/domManager.js":function(module,__webpack_exports__,__webpack_require__){"use strict";eval('__webpack_require__.r(__webpack_exports__);\nfunction _createForOfIteratorHelper(o) { if (typeof Symbol === "undefined" || o[Symbol.iterator] == null) { if (Array.isArray(o) || (o = _unsupportedIterableToArray(o))) { var i = 0; var F = function F() {}; return { s: F, n: function n() { if (i >= o.length) return { done: true }; return { done: false, value: o[i++] }; }, e: function e(_e) { throw _e; }, f: F }; } throw new TypeError("Invalid attempt to iterate non-iterable instance.\\nIn order to be iterable, non-array objects must have a [Symbol.iterator]() method."); } var it, normalCompletion = true, didErr = false, err; return { s: function s() { it = o[Symbol.iterator](); }, n: function n() { var step = it.next(); normalCompletion = step.done; return step; }, e: function e(_e2) { didErr = true; err = _e2; }, f: function f() { try { if (!normalCompletion && it["return"] != null) it["return"](); } finally { if (didErr) throw err; } } }; }\n\nfunction _unsupportedIterableToArray(o, minLen) { if (!o) return; if (typeof o === "string") return _arrayLikeToArray(o, minLen); var n = Object.prototype.toString.call(o).slice(8, -1); if (n === "Object" && o.constructor) n = o.constructor.name; if (n === "Map" || n === "Set") return Array.from(n); if (n === "Arguments" || /^(?:Ui|I)nt(?:8|16|32)(?:Clamped)?Array$/.test(n)) return _arrayLikeToArray(o, minLen); }\n\nfunction _arrayLikeToArray(arr, len) { if (len == null || len > arr.length) len = arr.length; for (var i = 0, arr2 = new Array(len); i < len; i++) { arr2[i] = arr[i]; } return arr2; }\n\nvar updateUserMessages = function updateUserMessages(userMessages) {\n  chat.innerHTML = \'\';\n\n  var _iterator = _createForOfIteratorHelper(userMessages),\n      _step;\n\n  try {\n    for (_iterator.s(); !(_step = _iterator.n()).done;) {\n      var message = _step.value;\n      var node = document.createElement("DIV");\n      var textnode = document.createTextNode(message);\n      node.appendChild(textnode);\n      chat.appendChild(node);\n    }\n  } catch (err) {\n    _iterator.e(err);\n  } finally {\n    _iterator.f();\n  }\n};\n\n/* harmony default export */ __webpack_exports__["default"] = ({\n  updateUserMessages: updateUserMessages\n});\n\n//# sourceURL=webpack:///./src/js/domManager.js?')},"./src/js/emitter.js":function(module,__webpack_exports__,__webpack_require__){"use strict";eval("__webpack_require__.r(__webpack_exports__);\nvar emitConnection = function emitConnection(pseudo) {\n  socket.emit(\"pseudo\", pseudo);\n};\n\nvar emitEraseAll = function emitEraseAll() {\n  socket.emit('eraseDrawing');\n};\n\nvar emitDrawingStart = function emitDrawingStart(mouse) {\n  socket.emit('startDrawing', mouse);\n};\n\nvar emitDrawingTrace = function emitDrawingTrace(mouse) {\n  socket.emit('traceDrawing', mouse);\n};\n\nvar emitNewMessage = function emitNewMessage(msg) {\n  socket.emit('message', msg);\n};\n\n/* harmony default export */ __webpack_exports__[\"default\"] = ({\n  emitConnection: emitConnection,\n  emitEraseAll: emitEraseAll,\n  emitDrawingStart: emitDrawingStart,\n  emitDrawingTrace: emitDrawingTrace,\n  emitNewMessage: emitNewMessage\n});\n\n//# sourceURL=webpack:///./src/js/emitter.js?")},"./src/js/game.js":function(module,__webpack_exports__,__webpack_require__){"use strict";eval('__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var _canvasManager_js__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ./canvasManager.js */ "./src/js/canvasManager.js");\n/* harmony import */ var _domManager_js__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ./domManager.js */ "./src/js/domManager.js");\nPromise.resolve(/*! import() */).then(__webpack_require__.bind(null, /*! ./connection.js */ "./src/js/connection.js"));\nPromise.resolve(/*! import() */).then(__webpack_require__.bind(null, /*! ./emitter.js */ "./src/js/emitter.js"));\n\n\nsocket.on(\'drawingStarted\', function (mouse) {\n  _canvasManager_js__WEBPACK_IMPORTED_MODULE_0__["default"].startDrawing(mouse);\n});\nsocket.on(\'drawingTraced\', function (mouse) {\n  _canvasManager_js__WEBPACK_IMPORTED_MODULE_0__["default"].trace(mouse);\n});\nsocket.on(\'drawingErased\', function (_) {\n  _canvasManager_js__WEBPACK_IMPORTED_MODULE_0__["default"].eraseDrawing();\n});\nsocket.on(\'listUsers\', function (list) {\n  console.log(list);\n});\nsocket.on(\'userMessages\', function (userMessages) {\n  _domManager_js__WEBPACK_IMPORTED_MODULE_1__["default"].updateUserMessages(userMessages);\n});\n\n//# sourceURL=webpack:///./src/js/game.js?')},"./src/js/tchat.js":function(module,__webpack_exports__,__webpack_require__){"use strict";eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var _emitter_js__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ./emitter.js */ \"./src/js/emitter.js\");\n\ntchatForm.addEventListener('submit', function (evt) {\n  evt.preventDefault();\n  _emitter_js__WEBPACK_IMPORTED_MODULE_0__[\"default\"].emitNewMessage(msg.value);\n  msg.value = '';\n});\n\n//# sourceURL=webpack:///./src/js/tchat.js?")},0:function(module,exports,__webpack_require__){eval('__webpack_require__(/*! ./src/js/game.js */"./src/js/game.js");\n__webpack_require__(/*! ./src/js/connection.js */"./src/js/connection.js");\n__webpack_require__(/*! ./src/js/tchat.js */"./src/js/tchat.js");\n__webpack_require__(/*! ./src/js/canvasManager.js */"./src/js/canvasManager.js");\nmodule.exports = __webpack_require__(/*! ./src/js/domManager.js */"./src/js/domManager.js");\n\n\n//# sourceURL=webpack:///multi_./src/js/game.js_./src/js/connection.js_./src/js/tchat.js_./src/js/canvasManager.js_./src/js/domManager.js?')}});
+!function(e){var n={};function t(r){if(n[r])return n[r].exports;var s=n[r]={i:r,l:!1,exports:{}};return e[r].call(s.exports,s,s.exports,t),s.l=!0,s.exports}t.m=e,t.c=n,t.d=function(e,n,r){t.o(e,n)||Object.defineProperty(e,n,{enumerable:!0,get:r})},t.r=function(e){"undefined"!=typeof Symbol&&Symbol.toStringTag&&Object.defineProperty(e,Symbol.toStringTag,{value:"Module"}),Object.defineProperty(e,"__esModule",{value:!0})},t.t=function(e,n){if(1&n&&(e=t(e)),8&n)return e;if(4&n&&"object"==typeof e&&e&&e.__esModule)return e;var r=Object.create(null);if(t.r(r),Object.defineProperty(r,"default",{enumerable:!0,value:e}),2&n&&"string"!=typeof e)for(var s in e)t.d(r,s,function(n){return e[n]}.bind(null,s));return r},t.n=function(e){var n=e&&e.__esModule?function(){return e.default}:function(){return e};return t.d(n,"a",n),n},t.o=function(e,n){return Object.prototype.hasOwnProperty.call(e,n)},t.p="",t(t.s=0)}({"./node_modules/events/events.js":function(module,exports,__webpack_require__){"use strict";eval("// Copyright Joyent, Inc. and other Node contributors.\n//\n// Permission is hereby granted, free of charge, to any person obtaining a\n// copy of this software and associated documentation files (the\n// \"Software\"), to deal in the Software without restriction, including\n// without limitation the rights to use, copy, modify, merge, publish,\n// distribute, sublicense, and/or sell copies of the Software, and to permit\n// persons to whom the Software is furnished to do so, subject to the\n// following conditions:\n//\n// The above copyright notice and this permission notice shall be included\n// in all copies or substantial portions of the Software.\n//\n// THE SOFTWARE IS PROVIDED \"AS IS\", WITHOUT WARRANTY OF ANY KIND, EXPRESS\n// OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF\n// MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN\n// NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM,\n// DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR\n// OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE\n// USE OR OTHER DEALINGS IN THE SOFTWARE.\n\n\n\nvar R = typeof Reflect === 'object' ? Reflect : null\nvar ReflectApply = R && typeof R.apply === 'function'\n  ? R.apply\n  : function ReflectApply(target, receiver, args) {\n    return Function.prototype.apply.call(target, receiver, args);\n  }\n\nvar ReflectOwnKeys\nif (R && typeof R.ownKeys === 'function') {\n  ReflectOwnKeys = R.ownKeys\n} else if (Object.getOwnPropertySymbols) {\n  ReflectOwnKeys = function ReflectOwnKeys(target) {\n    return Object.getOwnPropertyNames(target)\n      .concat(Object.getOwnPropertySymbols(target));\n  };\n} else {\n  ReflectOwnKeys = function ReflectOwnKeys(target) {\n    return Object.getOwnPropertyNames(target);\n  };\n}\n\nfunction ProcessEmitWarning(warning) {\n  if (console && console.warn) console.warn(warning);\n}\n\nvar NumberIsNaN = Number.isNaN || function NumberIsNaN(value) {\n  return value !== value;\n}\n\nfunction EventEmitter() {\n  EventEmitter.init.call(this);\n}\nmodule.exports = EventEmitter;\n\n// Backwards-compat with node 0.10.x\nEventEmitter.EventEmitter = EventEmitter;\n\nEventEmitter.prototype._events = undefined;\nEventEmitter.prototype._eventsCount = 0;\nEventEmitter.prototype._maxListeners = undefined;\n\n// By default EventEmitters will print a warning if more than 10 listeners are\n// added to it. This is a useful default which helps finding memory leaks.\nvar defaultMaxListeners = 10;\n\nfunction checkListener(listener) {\n  if (typeof listener !== 'function') {\n    throw new TypeError('The \"listener\" argument must be of type Function. Received type ' + typeof listener);\n  }\n}\n\nObject.defineProperty(EventEmitter, 'defaultMaxListeners', {\n  enumerable: true,\n  get: function() {\n    return defaultMaxListeners;\n  },\n  set: function(arg) {\n    if (typeof arg !== 'number' || arg < 0 || NumberIsNaN(arg)) {\n      throw new RangeError('The value of \"defaultMaxListeners\" is out of range. It must be a non-negative number. Received ' + arg + '.');\n    }\n    defaultMaxListeners = arg;\n  }\n});\n\nEventEmitter.init = function() {\n\n  if (this._events === undefined ||\n      this._events === Object.getPrototypeOf(this)._events) {\n    this._events = Object.create(null);\n    this._eventsCount = 0;\n  }\n\n  this._maxListeners = this._maxListeners || undefined;\n};\n\n// Obviously not all Emitters should be limited to 10. This function allows\n// that to be increased. Set to zero for unlimited.\nEventEmitter.prototype.setMaxListeners = function setMaxListeners(n) {\n  if (typeof n !== 'number' || n < 0 || NumberIsNaN(n)) {\n    throw new RangeError('The value of \"n\" is out of range. It must be a non-negative number. Received ' + n + '.');\n  }\n  this._maxListeners = n;\n  return this;\n};\n\nfunction _getMaxListeners(that) {\n  if (that._maxListeners === undefined)\n    return EventEmitter.defaultMaxListeners;\n  return that._maxListeners;\n}\n\nEventEmitter.prototype.getMaxListeners = function getMaxListeners() {\n  return _getMaxListeners(this);\n};\n\nEventEmitter.prototype.emit = function emit(type) {\n  var args = [];\n  for (var i = 1; i < arguments.length; i++) args.push(arguments[i]);\n  var doError = (type === 'error');\n\n  var events = this._events;\n  if (events !== undefined)\n    doError = (doError && events.error === undefined);\n  else if (!doError)\n    return false;\n\n  // If there is no 'error' event listener then throw.\n  if (doError) {\n    var er;\n    if (args.length > 0)\n      er = args[0];\n    if (er instanceof Error) {\n      // Note: The comments on the `throw` lines are intentional, they show\n      // up in Node's output if this results in an unhandled exception.\n      throw er; // Unhandled 'error' event\n    }\n    // At least give some kind of context to the user\n    var err = new Error('Unhandled error.' + (er ? ' (' + er.message + ')' : ''));\n    err.context = er;\n    throw err; // Unhandled 'error' event\n  }\n\n  var handler = events[type];\n\n  if (handler === undefined)\n    return false;\n\n  if (typeof handler === 'function') {\n    ReflectApply(handler, this, args);\n  } else {\n    var len = handler.length;\n    var listeners = arrayClone(handler, len);\n    for (var i = 0; i < len; ++i)\n      ReflectApply(listeners[i], this, args);\n  }\n\n  return true;\n};\n\nfunction _addListener(target, type, listener, prepend) {\n  var m;\n  var events;\n  var existing;\n\n  checkListener(listener);\n\n  events = target._events;\n  if (events === undefined) {\n    events = target._events = Object.create(null);\n    target._eventsCount = 0;\n  } else {\n    // To avoid recursion in the case that type === \"newListener\"! Before\n    // adding it to the listeners, first emit \"newListener\".\n    if (events.newListener !== undefined) {\n      target.emit('newListener', type,\n                  listener.listener ? listener.listener : listener);\n\n      // Re-assign `events` because a newListener handler could have caused the\n      // this._events to be assigned to a new object\n      events = target._events;\n    }\n    existing = events[type];\n  }\n\n  if (existing === undefined) {\n    // Optimize the case of one listener. Don't need the extra array object.\n    existing = events[type] = listener;\n    ++target._eventsCount;\n  } else {\n    if (typeof existing === 'function') {\n      // Adding the second element, need to change to array.\n      existing = events[type] =\n        prepend ? [listener, existing] : [existing, listener];\n      // If we've already got an array, just append.\n    } else if (prepend) {\n      existing.unshift(listener);\n    } else {\n      existing.push(listener);\n    }\n\n    // Check for listener leak\n    m = _getMaxListeners(target);\n    if (m > 0 && existing.length > m && !existing.warned) {\n      existing.warned = true;\n      // No error code for this since it is a Warning\n      // eslint-disable-next-line no-restricted-syntax\n      var w = new Error('Possible EventEmitter memory leak detected. ' +\n                          existing.length + ' ' + String(type) + ' listeners ' +\n                          'added. Use emitter.setMaxListeners() to ' +\n                          'increase limit');\n      w.name = 'MaxListenersExceededWarning';\n      w.emitter = target;\n      w.type = type;\n      w.count = existing.length;\n      ProcessEmitWarning(w);\n    }\n  }\n\n  return target;\n}\n\nEventEmitter.prototype.addListener = function addListener(type, listener) {\n  return _addListener(this, type, listener, false);\n};\n\nEventEmitter.prototype.on = EventEmitter.prototype.addListener;\n\nEventEmitter.prototype.prependListener =\n    function prependListener(type, listener) {\n      return _addListener(this, type, listener, true);\n    };\n\nfunction onceWrapper() {\n  if (!this.fired) {\n    this.target.removeListener(this.type, this.wrapFn);\n    this.fired = true;\n    if (arguments.length === 0)\n      return this.listener.call(this.target);\n    return this.listener.apply(this.target, arguments);\n  }\n}\n\nfunction _onceWrap(target, type, listener) {\n  var state = { fired: false, wrapFn: undefined, target: target, type: type, listener: listener };\n  var wrapped = onceWrapper.bind(state);\n  wrapped.listener = listener;\n  state.wrapFn = wrapped;\n  return wrapped;\n}\n\nEventEmitter.prototype.once = function once(type, listener) {\n  checkListener(listener);\n  this.on(type, _onceWrap(this, type, listener));\n  return this;\n};\n\nEventEmitter.prototype.prependOnceListener =\n    function prependOnceListener(type, listener) {\n      checkListener(listener);\n      this.prependListener(type, _onceWrap(this, type, listener));\n      return this;\n    };\n\n// Emits a 'removeListener' event if and only if the listener was removed.\nEventEmitter.prototype.removeListener =\n    function removeListener(type, listener) {\n      var list, events, position, i, originalListener;\n\n      checkListener(listener);\n\n      events = this._events;\n      if (events === undefined)\n        return this;\n\n      list = events[type];\n      if (list === undefined)\n        return this;\n\n      if (list === listener || list.listener === listener) {\n        if (--this._eventsCount === 0)\n          this._events = Object.create(null);\n        else {\n          delete events[type];\n          if (events.removeListener)\n            this.emit('removeListener', type, list.listener || listener);\n        }\n      } else if (typeof list !== 'function') {\n        position = -1;\n\n        for (i = list.length - 1; i >= 0; i--) {\n          if (list[i] === listener || list[i].listener === listener) {\n            originalListener = list[i].listener;\n            position = i;\n            break;\n          }\n        }\n\n        if (position < 0)\n          return this;\n\n        if (position === 0)\n          list.shift();\n        else {\n          spliceOne(list, position);\n        }\n\n        if (list.length === 1)\n          events[type] = list[0];\n\n        if (events.removeListener !== undefined)\n          this.emit('removeListener', type, originalListener || listener);\n      }\n\n      return this;\n    };\n\nEventEmitter.prototype.off = EventEmitter.prototype.removeListener;\n\nEventEmitter.prototype.removeAllListeners =\n    function removeAllListeners(type) {\n      var listeners, events, i;\n\n      events = this._events;\n      if (events === undefined)\n        return this;\n\n      // not listening for removeListener, no need to emit\n      if (events.removeListener === undefined) {\n        if (arguments.length === 0) {\n          this._events = Object.create(null);\n          this._eventsCount = 0;\n        } else if (events[type] !== undefined) {\n          if (--this._eventsCount === 0)\n            this._events = Object.create(null);\n          else\n            delete events[type];\n        }\n        return this;\n      }\n\n      // emit removeListener for all listeners on all events\n      if (arguments.length === 0) {\n        var keys = Object.keys(events);\n        var key;\n        for (i = 0; i < keys.length; ++i) {\n          key = keys[i];\n          if (key === 'removeListener') continue;\n          this.removeAllListeners(key);\n        }\n        this.removeAllListeners('removeListener');\n        this._events = Object.create(null);\n        this._eventsCount = 0;\n        return this;\n      }\n\n      listeners = events[type];\n\n      if (typeof listeners === 'function') {\n        this.removeListener(type, listeners);\n      } else if (listeners !== undefined) {\n        // LIFO order\n        for (i = listeners.length - 1; i >= 0; i--) {\n          this.removeListener(type, listeners[i]);\n        }\n      }\n\n      return this;\n    };\n\nfunction _listeners(target, type, unwrap) {\n  var events = target._events;\n\n  if (events === undefined)\n    return [];\n\n  var evlistener = events[type];\n  if (evlistener === undefined)\n    return [];\n\n  if (typeof evlistener === 'function')\n    return unwrap ? [evlistener.listener || evlistener] : [evlistener];\n\n  return unwrap ?\n    unwrapListeners(evlistener) : arrayClone(evlistener, evlistener.length);\n}\n\nEventEmitter.prototype.listeners = function listeners(type) {\n  return _listeners(this, type, true);\n};\n\nEventEmitter.prototype.rawListeners = function rawListeners(type) {\n  return _listeners(this, type, false);\n};\n\nEventEmitter.listenerCount = function(emitter, type) {\n  if (typeof emitter.listenerCount === 'function') {\n    return emitter.listenerCount(type);\n  } else {\n    return listenerCount.call(emitter, type);\n  }\n};\n\nEventEmitter.prototype.listenerCount = listenerCount;\nfunction listenerCount(type) {\n  var events = this._events;\n\n  if (events !== undefined) {\n    var evlistener = events[type];\n\n    if (typeof evlistener === 'function') {\n      return 1;\n    } else if (evlistener !== undefined) {\n      return evlistener.length;\n    }\n  }\n\n  return 0;\n}\n\nEventEmitter.prototype.eventNames = function eventNames() {\n  return this._eventsCount > 0 ? ReflectOwnKeys(this._events) : [];\n};\n\nfunction arrayClone(arr, n) {\n  var copy = new Array(n);\n  for (var i = 0; i < n; ++i)\n    copy[i] = arr[i];\n  return copy;\n}\n\nfunction spliceOne(list, index) {\n  for (; index + 1 < list.length; index++)\n    list[index] = list[index + 1];\n  list.pop();\n}\n\nfunction unwrapListeners(arr) {\n  var ret = new Array(arr.length);\n  for (var i = 0; i < ret.length; ++i) {\n    ret[i] = arr[i].listener || arr[i];\n  }\n  return ret;\n}\n\n\n//# sourceURL=webpack:///./node_modules/events/events.js?")},"./src/js/canvasManager.js":function(module,__webpack_exports__,__webpack_require__){"use strict";eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var _webSocket_js__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ./webSocket.js */ \"./src/js/webSocket.js\");\n //Create canvas\n\nvar canvas = document.getElementById('myCanvas');\nvar context = canvas.getContext('2d');\nvar mouse = {\n  x: 0,\n  y: 0,\n  width: 1,\n  color: \"black\"\n}; //Set background\n\ncontext.fillStyle = \"white\";\ncontext.fillRect(0, 0, 700, 500); //Canvas controls behaviour\n\neraseAll.addEventListener('click', function (evt) {\n  evt.preventDefault();\n  eraseDrawing();\n  _webSocket_js__WEBPACK_IMPORTED_MODULE_0__[\"default\"].emitEraseAll();\n});\n\nvar eraseDrawing = function eraseDrawing() {\n  context.clearRect(0, 0, canvas.width, canvas.height);\n};\n\npencil.addEventListener('click', function (evt) {\n  evt.preventDefault();\n  mouse.width = 1;\n  mouse.color = \"black\";\n});\neraser.addEventListener('click', function (evt) {\n  evt.preventDefault();\n  mouse.width = 100;\n  mouse.color = \"white\";\n}); // canvas behaviour\n//drawing start\n\ncanvas.addEventListener('mousedown', function (evt) {\n  evt.preventDefault();\n  getMousePos(evt);\n  startDrawing(mouse);\n  _webSocket_js__WEBPACK_IMPORTED_MODULE_0__[\"default\"].emitDrawingStart(mouse);\n  startTracing();\n});\n\nvar startDrawing = function startDrawing(currentMouse) {\n  context.lineWidth = currentMouse.width;\n  context.strokeStyle = currentMouse.color;\n  context.beginPath();\n  context.moveTo(currentMouse.x, currentMouse.y);\n}; // drawing tracing\n\n\nvar startTracing = function startTracing() {\n  canvas.addEventListener('mousemove', mouseMove, false);\n};\n\nvar mouseMove = function mouseMove(evt) {\n  getMousePos(evt);\n  trace(mouse);\n  _webSocket_js__WEBPACK_IMPORTED_MODULE_0__[\"default\"].emitDrawingTrace(mouse);\n};\n\nvar trace = function trace(currentMouse) {\n  context.lineTo(currentMouse.x, currentMouse.y);\n  context.stroke();\n}; //stopping drawing\n\n\ncanvas.addEventListener('mouseup', function (evt) {\n  stopDrawing();\n}, false);\n\nvar stopDrawing = function stopDrawing() {\n  canvas.removeEventListener('mousemove', mouseMove, false);\n};\n\nvar getMousePos = function getMousePos(evt) {\n  var rect = canvas.getBoundingClientRect();\n  mouse.x = evt.clientX - rect.left;\n  mouse.y = evt.clientY - rect.top;\n};\n\n/* harmony default export */ __webpack_exports__[\"default\"] = ({\n  startDrawing: startDrawing,\n  trace: trace,\n  stopDrawing: stopDrawing,\n  eraseDrawing: eraseDrawing\n});\n\n//# sourceURL=webpack:///./src/js/canvasManager.js?")},"./src/js/connection.js":function(module,__webpack_exports__,__webpack_require__){"use strict";eval('__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var _webSocket_js__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ./webSocket.js */ "./src/js/webSocket.js");\n //Get the pseudo of user\n\nvar pseudo = prompt("Entrer votre pseudonyme:", "");\n_webSocket_js__WEBPACK_IMPORTED_MODULE_0__["default"].emitConnection(pseudo);\n\n//# sourceURL=webpack:///./src/js/connection.js?')},"./src/js/domManager.js":function(module,__webpack_exports__,__webpack_require__){"use strict";eval('__webpack_require__.r(__webpack_exports__);\nfunction _createForOfIteratorHelper(o) { if (typeof Symbol === "undefined" || o[Symbol.iterator] == null) { if (Array.isArray(o) || (o = _unsupportedIterableToArray(o))) { var i = 0; var F = function F() {}; return { s: F, n: function n() { if (i >= o.length) return { done: true }; return { done: false, value: o[i++] }; }, e: function e(_e) { throw _e; }, f: F }; } throw new TypeError("Invalid attempt to iterate non-iterable instance.\\nIn order to be iterable, non-array objects must have a [Symbol.iterator]() method."); } var it, normalCompletion = true, didErr = false, err; return { s: function s() { it = o[Symbol.iterator](); }, n: function n() { var step = it.next(); normalCompletion = step.done; return step; }, e: function e(_e2) { didErr = true; err = _e2; }, f: function f() { try { if (!normalCompletion && it["return"] != null) it["return"](); } finally { if (didErr) throw err; } } }; }\n\nfunction _unsupportedIterableToArray(o, minLen) { if (!o) return; if (typeof o === "string") return _arrayLikeToArray(o, minLen); var n = Object.prototype.toString.call(o).slice(8, -1); if (n === "Object" && o.constructor) n = o.constructor.name; if (n === "Map" || n === "Set") return Array.from(n); if (n === "Arguments" || /^(?:Ui|I)nt(?:8|16|32)(?:Clamped)?Array$/.test(n)) return _arrayLikeToArray(o, minLen); }\n\nfunction _arrayLikeToArray(arr, len) { if (len == null || len > arr.length) len = arr.length; for (var i = 0, arr2 = new Array(len); i < len; i++) { arr2[i] = arr[i]; } return arr2; }\n\nvar updateUserMessages = function updateUserMessages(userMessages) {\n  chat.innerHTML = \'\';\n\n  var _iterator = _createForOfIteratorHelper(userMessages),\n      _step;\n\n  try {\n    for (_iterator.s(); !(_step = _iterator.n()).done;) {\n      var message = _step.value;\n      var node = document.createElement("DIV");\n      var textnode = document.createTextNode(message);\n      node.appendChild(textnode);\n      chat.appendChild(node);\n    }\n  } catch (err) {\n    _iterator.e(err);\n  } finally {\n    _iterator.f();\n  }\n};\n\n/* harmony default export */ __webpack_exports__["default"] = ({\n  updateUserMessages: updateUserMessages\n});\n\n//# sourceURL=webpack:///./src/js/domManager.js?')},"./src/js/game.js":function(module,__webpack_exports__,__webpack_require__){"use strict";eval('__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var _canvasManager_js__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ./canvasManager.js */ "./src/js/canvasManager.js");\n/* harmony import */ var _domManager_js__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ./domManager.js */ "./src/js/domManager.js");\n/* harmony import */ var _webSocket_js__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ./webSocket.js */ "./src/js/webSocket.js");\nPromise.resolve(/*! import() */).then(__webpack_require__.bind(null, /*! ./connection.js */ "./src/js/connection.js"));\n\n\n\nvar emitter = _webSocket_js__WEBPACK_IMPORTED_MODULE_2__["default"].emitter;\nemitter.on(\'drawingStarted\', function (mouse) {\n  _canvasManager_js__WEBPACK_IMPORTED_MODULE_0__["default"].startDrawing(mouse);\n});\nemitter.on(\'drawingTraced\', function (mouse) {\n  _canvasManager_js__WEBPACK_IMPORTED_MODULE_0__["default"].trace(mouse);\n});\nemitter.on(\'drawingErased\', function (_) {\n  _canvasManager_js__WEBPACK_IMPORTED_MODULE_0__["default"].eraseDrawing();\n});\nemitter.on(\'listUsers\', function (list) {\n  console.log(list);\n});\nemitter.on(\'userMessages\', function (userMessages) {\n  _domManager_js__WEBPACK_IMPORTED_MODULE_1__["default"].updateUserMessages(userMessages);\n});\n\n//# sourceURL=webpack:///./src/js/game.js?')},"./src/js/tchat.js":function(module,__webpack_exports__,__webpack_require__){"use strict";eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var _webSocket_js__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ./webSocket.js */ \"./src/js/webSocket.js\");\n\ntchatForm.addEventListener('submit', function (evt) {\n  evt.preventDefault();\n  _webSocket_js__WEBPACK_IMPORTED_MODULE_0__[\"default\"].emitNewMessage(msg.value);\n  msg.value = '';\n});\n\n//# sourceURL=webpack:///./src/js/tchat.js?")},"./src/js/webSocket.js":function(module,__webpack_exports__,__webpack_require__){"use strict";eval("__webpack_require__.r(__webpack_exports__);\nvar EventEmitter = __webpack_require__(/*! events */ \"./node_modules/events/events.js\");\n\nvar emitter = new EventEmitter();\n\nvar emitDrawingStart = function emitDrawingStart(mouse) {\n  socket.emit('startDrawing', mouse);\n};\n\nsocket.on('drawingStarted', function (mouse) {\n  emitter.emit('drawingStarted', mouse);\n});\n\nvar emitDrawingTrace = function emitDrawingTrace(mouse) {\n  socket.emit('traceDrawing', mouse);\n};\n\nsocket.on('drawingTraced', function (mouse) {\n  emitter.emit('drawingTraced', mouse);\n});\n\nvar emitEraseAll = function emitEraseAll() {\n  socket.emit('eraseDrawing');\n};\n\nsocket.on('drawingErased', function (_) {\n  emitter.emit('drawingErased');\n});\n\nvar emitConnection = function emitConnection(pseudo) {\n  socket.emit(\"pseudo\", pseudo);\n};\n\nsocket.on('listUsers', function (list) {\n  emitter.emit('listUsers', list);\n});\n\nvar emitNewMessage = function emitNewMessage(msg) {\n  socket.emit('message', msg);\n};\n\nsocket.on('userMessages', function (userMessages) {\n  emitter.emit('userMessages', userMessages);\n});\n/* harmony default export */ __webpack_exports__[\"default\"] = ({\n  emitter: emitter,\n  EventEmitter: EventEmitter,\n  emitConnection: emitConnection,\n  emitEraseAll: emitEraseAll,\n  emitDrawingStart: emitDrawingStart,\n  emitDrawingTrace: emitDrawingTrace,\n  emitNewMessage: emitNewMessage\n});\n\n//# sourceURL=webpack:///./src/js/webSocket.js?")},0:function(module,exports,__webpack_require__){eval('__webpack_require__(/*! ./src/js/game.js */"./src/js/game.js");\n__webpack_require__(/*! ./src/js/connection.js */"./src/js/connection.js");\n__webpack_require__(/*! ./src/js/tchat.js */"./src/js/tchat.js");\n__webpack_require__(/*! ./src/js/canvasManager.js */"./src/js/canvasManager.js");\n__webpack_require__(/*! ./src/js/domManager.js */"./src/js/domManager.js");\nmodule.exports = __webpack_require__(/*! ./src/js/webSocket.js */"./src/js/webSocket.js");\n\n\n//# sourceURL=webpack:///multi_./src/js/game.js_./src/js/connection.js_./src/js/tchat.js_./src/js/canvasManager.js_./src/js/domManager.js_./src/js/webSocket.js?')}});
