+++ public/js/game.js.LICENSE.txt
@@ -12,9 +12,9 @@
   !*** ./src/js/tchat.js ***!
   \*************************/
 
-/*!***************************!*\
-  !*** ./src/js/emitter.js ***!
-  \***************************/
+/*!*****************************!*\
+  !*** ./src/js/webSocket.js ***!
+  \*****************************/
 
 /*!******************************!*\
   !*** ./src/js/connection.js ***!
@@ -28,6 +28,10 @@
   !*** ./src/js/canvasManager.js ***!
   \*********************************/
 
-/*!************************************************************************************************************************!*\
-  !*** multi ./src/js/game.js ./src/js/connection.js ./src/js/tchat.js ./src/js/canvasManager.js ./src/js/domManager.js ***!
-  \************************************************************************************************************************/
+/*!***************************************!*\
+  !*** ./node_modules/events/events.js ***!
+  \***************************************/
+
+/*!**********************************************************************************************************************************************!*\
+  !*** multi ./src/js/game.js ./src/js/connection.js ./src/js/tchat.js ./src/js/canvasManager.js ./src/js/domManager.js ./src/js/webSocket.js ***!
+  \**********************************************************************************************************************************************/
