+++ src/js/connection.js
@@ -1,8 +1,6 @@
-import emitter from './emitter.js';
+import emitter from './webSocket.js';
 
 //Get the pseudo of user
 const pseudo = prompt("Entrer votre pseudonyme:","");
 
 emitter.emitConnection(pseudo);
-    
-
