+++ src/js/tchat.js
@@ -1,5 +1,4 @@
-import emitter from './emitter.js';
-
+import emitter from './webSocket.js';
 
 tchatForm.addEventListener('submit', function(evt){
     evt.preventDefault();
